+++ src/project/rollup-plugin-output-css.ts
@@ -2,7 +2,7 @@ import {Plugin} from 'rollup';
 import {dirname, join, relative} from 'path';
 import sourcemapCodec from 'sourcemap-codec';
 
-import {blue, gray} from '../colors/terminal.js';
+import {blue, gray} from '../utils/terminal.js';
 import {outputFile} from '../fs/nodeFs.js';
 import {SystemLogger, Logger} from '../utils/log.js';
 import {GroCssBuild, GroCssBundle} from './types.js';