+++ src/project/svelte-preprocess-typescript.ts
@@ -1,7 +1,7 @@
 import ts from 'typescript';
 import {PreprocessorGroup} from 'svelte/types/compiler/preprocess';
 
-import {magenta, red} from '../colors/terminal.js';
+import {magenta, red} from '../utils/terminal.js';
 import {SystemLogger, Logger} from '../utils/log.js';
 import {loadTsconfig} from '../build/tsBuildHelpers.js';
 import {printPath} from '../utils/print.js';